requirements.txt
torch
transformers==4.40.0
tokenizers==0.19.1
peft==0.12.0
lm-eval==0.4.7
datasets==2.18.0
ml_collections
einops
omegaconf
pytorch-minimize